+++ src/server/visibility/filters_mask.rs
@@ -3,10 +3,7 @@ use core::iter;
 use bevy::prelude::*;
 
 use super::registry::FilterRegistry;
-use crate::shared::replication::{
-    registry::{ComponentIndex, component_mask::ComponentMask},
-    visibility::VisibilityScope,
-};
+use crate::shared::replication::{registry::ComponentIndex, visibility::VisibilityScope};
 
 /// Bitset of visibility filters for an entity for a client.
 ///
@@ -62,19 +59,7 @@ impl FiltersMask {
         self.iter().any(|bit| match registry.scope(bit) {
             VisibilityScope::Entity => true,
             VisibilityScope::Components(component_mask) => component_mask.contains(index),
-        })
-    }
-
-    /// Returns an iterator over hidden components for an entity.
-    pub(crate) fn hidden_components(
-        self,
-        registry: &FilterRegistry,
-    ) -> impl Iterator<Item = &ComponentMask> {
-        self.iter().map(|bit| match registry.scope(bit) {
-            VisibilityScope::Entity => {
-                panic!("if the entity is hidden, iteration over hidden components shouldn't happen")
-            }
-            VisibilityScope::Components(component_mask) => component_mask,
+            VisibilityScope::OnlyComponents(component_mask) => !component_mask.contains(index),
         })
     }
 }